ml4pl: Overhaul graph representation.
This patch combines a number of improvements into a single atomic
change:

(1) The program graph proto is restructured to include a flat table of
strings, rather than per-node/node-graph strings. Proto fields which
previously contained strings now contain indices into this strings
table. This has the effect of normalizing the graph representation by
enabling the same string to be re-used across multiple elements in a
graph.

(2) Re-implement random graph generator in C++. This unifies the C++
and Python "graph builder" classes and provides a single C++
implementation with pybind11 bindings for python.

(3) Replace the old test set of program graphs derived from LLVM IR in
favor of a test fixture which generates the program graphs from LLVM
IR at runtime. This means that any changes to the graph construction
will be reflected in all tests.

github.com/ChrisCummins/ProGraML/issues/55

+++ deeplearning/ml4pl/graphs/xla2graph/hlo_module_graph_builder.h
@@ -42,17 +42,17 @@ class HloModuleGraphBuilder : GraphBuilder {
   labm8::StatusOr<FunctionEntryExits> VisitComputation(
       const xla::HloComputationProto& computation);
 
-  labm8::StatusOr<size_t> VisitInstruction(
-      const xla::HloInstructionProto& instruction, size_t functionNumber,
-      size_t entryInstruction);
+  labm8::StatusOr<int> VisitInstruction(
+      const xla::HloInstructionProto& instruction, int functionNumber,
+      int entryInstruction);
 
  private:
   // A map from computations to their entry and exit nodes.
   absl::flat_hash_map<labm8::int64, FunctionEntryExits> computations_;
   // A map of instruction IDs to their node number.
-  absl::flat_hash_map<labm8::int64, size_t> instructions_;
+  absl::flat_hash_map<labm8::int64, int> instructions_;
   // A map of instruction IDs to the data element produced by the instruction.
-  absl::flat_hash_map<labm8::int64, size_t> producers_;
+  absl::flat_hash_map<labm8::int64, int> producers_;
 };
 
 }  // namespace ml4pl
